edit models and add change email feature. refactor a bit
