+++ add_default_user.py
@@ -9,10 +9,11 @@ load_dotenv()
 app = create_app()
 with app.app_context():
     password_plain = os.environ.get("ADMIN_PASSWORD")
+    email = os.environ.get("ADMIN_EMAIL")
     if not password_plain:
         raise ValueError("ADMIN_PASSWORD environment variable not set.")
     hashed_password = generate_password_hash(password_plain)
-    new_user = User(username="admin", password_hash=hashed_password)
+    new_user = User(username="admin", email=email, password_hash=hashed_password)
     db.session.add(new_user)
     db.session.commit()
     print("User added successfully!")
