+++ app/templates/settings.html
@@ -1,105 +1,8 @@
-{% extends 'layout.html' %}
-{% block main_content %}
-<div class="flex items-center justify-center w-full h-full">
-  <form id="settingsForm" method="POST" class="flex flex-col items-center justify-center" style="width:500px;background:#fff;padding:32px 24px;gap:24px;border-radius:8px;box-shadow:0 2px 8px rgba(0,0,0,0.04);">
-    {{ form.hidden_tag() }}
-    <div style="width:344px;">
-      <label for="old_password" style="font-family:'Roboto';font-size:16px;color:rgba(0,0,0,0.87);margin-bottom:4px;">Old Password</label>
-      <div style="position:relative;">
-        {{ form.old_password(class="form-control", id="old_password", style="width:330px;height:54px;padding-right:40px;border:1px solid #6200EE;border-radius:4px;font-size:16px;color:rgba(0,0,0,0.87);") }}
-        <span style="position:absolute;right:12px;top:16px;cursor:pointer;" onclick="togglePassword('old_password', this)">
-          <svg id="eye_old_password" width="24" height="24" fill="none"><circle cx="12" cy="12" r="10" stroke="#6200EE" stroke-width="2"/><circle cx="12" cy="12" r="4" stroke="#6200EE" stroke-width="2"/></svg>
-        </span>
-      </div>
-    </div>
-    <div style="width:344px;">
-      <label for="new_password" style="font-family:'Roboto';font-size:16px;color:rgba(0,0,0,0.87);margin-bottom:4px;">New Password</label>
-      <div style="position:relative;">
-        {{ form.new_password(class="form-control", id="new_password", style="width:330px;height:54px;padding-right:40px;border:1px solid #6200EE;border-radius:4px;font-size:16px;color:rgba(0,0,0,0.87);") }}
-        <span style="position:absolute;right:12px;top:16px;cursor:pointer;" onclick="togglePassword('new_password', this)">
-          <svg id="eye_new_password" width="24" height="24" fill="none"><circle cx="12" cy="12" r="10" stroke="#6200EE" stroke-width="2"/><circle cx="12" cy="12" r="4" stroke="#6200EE" stroke-width="2"/></svg>
-        </span>
-      </div>
-    </div>
-    <div style="width:344px;">
-      <label for="repeat_new_password" style="font-family:'Roboto';font-size:16px;color:rgba(0,0,0,0.87);margin-bottom:4px;">Repeat New Password</label>
-      <div style="position:relative;">
-        {{ form.repeat_new_password(class="form-control", id="repeat_new_password", style="width:330px;height:54px;padding-right:40px;border:1px solid #6200EE;border-radius:4px;font-size:16px;color:rgba(0,0,0,0.87);") }}
-        <span style="position:absolute;right:12px;top:16px;cursor:pointer;" onclick="togglePassword('repeat_new_password', this)">
-          <svg id="eye_repeat_new_password" width="24" height="24" fill="none"><circle cx="12" cy="12" r="10" stroke="#6200EE" stroke-width="2"/><circle cx="12" cy="12" r="4" stroke="#6200EE" stroke-width="2"/></svg>
-        </span>
-      </div>
-    </div>
-    <div style="width:344px;position:relative;">
-      <div id="passwordRules" style="width:330px;height:54px;border:1px solid #6155F5;border-radius:4px;display:flex;align-items:center;justify-content:center;color:#6200EE;font-size:16px;background:#fff;position:relative;cursor:pointer;">
-        Password rules
-        <div id="rulesTooltip" style="display:none;position:absolute;top:60px;left:0;width:330px;background:#fff;border:1px solid #6200EE;border-radius:4px;padding:12px;color:#333;z-index:10;font-size:14px;box-shadow:0 2px 8px rgba(98,0,238,0.08);">
-          Password must be at least 8 characters long and contain letters, numbers, and special characters (e.g. !@#$%^&*()_+ etc.)
-        </div>
-      </div>
-    </div>
-    <div id="passwordError" style="color:red;font-size:14px;display:none;width:330px;text-align:center;"></div>
-    {{ form.submit(class="button cursor-pointer w-24", style="background:#6200EE;color:#fff;font-weight:500;font-size:14px;border-radius:4px;margin-top:16px;") }}
-  </form>
+{% extends 'layout.html' %} {% block main_content %}
+<div class="flex flex-col items-center justify-center w-full h-full gap-8">
+  <a href="{{ url_for('auth.change_password') }}" class="button"
+    >Change Password</a
+  >
+  <a href="{{ url_for('auth.change_email') }}" class="button">Change Email</a>
 </div>
-<script>
-function togglePassword(inputId, iconSpan) {
-  const input = document.getElementById(inputId);
-  if (input.type === "password") {
-    input.type = "text";
-    iconSpan.querySelector("svg").style.opacity = 0.5;
-  } else {
-    input.type = "password";
-    iconSpan.querySelector("svg").style.opacity = 1;
-  }
-}
-
-// Password rule validation
-function validatePasswordRule(pwd) {
-  // At least 8 characters, including letters, numbers, and special characters
-  return /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?]).{8,}$/.test(pwd);
-}
-
-// Password rules hover tooltip
-const rulesDiv = document.getElementById('passwordRules');
-const tooltip = document.getElementById('rulesTooltip');
-rulesDiv.addEventListener('mouseenter', () => { tooltip.style.display = 'block'; });
-rulesDiv.addEventListener('mouseleave', () => { tooltip.style.display = 'none'; });
-
-// Frontend validation
-document.getElementById('settingsForm').addEventListener('submit', function(e) {
-  const oldPwd = document.getElementById('old_password').value;
-  const newPwd = document.getElementById('new_password').value;
-  const repeatPwd = document.getElementById('repeat_new_password').value;
-  const errorDiv = document.getElementById('passwordError');
-
-  // New password rule validation
-  if (!validatePasswordRule(newPwd)) {
-    e.preventDefault();
-    errorDiv.textContent = "The new password must be at least 8 characters long and contain letters, numbers, and special characters.";
-    errorDiv.style.display = "block";
-    return;
-  }
-  // Consistency check for new password and confirmation
-  if (newPwd !== repeatPwd) {
-    e.preventDefault();
-    errorDiv.textContent = "The new password and confirmation password do not match. Please re-enter.";
-    errorDiv.style.display = "block";
-    return;
-  }
-  errorDiv.style.display = "none";
-});
-
-// Show prompt if new password does not meet the rule while typing
-document.getElementById('new_password').addEventListener('input', function() {
-  const newPwd = this.value;
-  const errorDiv = document.getElementById('passwordError');
-  if (newPwd && !validatePasswordRule(newPwd)) {
-    errorDiv.textContent = "The new password must be at least 8 characters long and contain letters, numbers, and special characters.";
-    errorDiv.style.display = "block";
-  } else {
-    errorDiv.style.display = "none";
-  }
-});
-</script>
 {% endblock %}
